split all examples
